@@ -60,14 +60,14 @@
     "uuid": "8.3.2"
   },
   "peerDependencies": {
-    "@material-ui/core": "4.11.0",
-    "@material-ui/icons": "4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.58",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.83",
+    "@mui/material": "^5.8.1",
     "ag-grid-community": "27.3.0",
     "ag-grid-react": "27.3.0",
     "i18next": "21.3.1",
-    "react": "^16.13.1",
-    "react-dom": "16.13.1"
+    "react": "18.1.0",
+    "react-dom": "18.1.0"
   },
   "devDependencies": {
     "@babel/plugin-proposal-optional-chaining": "7.16.7",
@@ -75,9 +75,9 @@
     "@babel/preset-typescript": "7.16.7",
     "@commitlint/cli": "17.0.0",
     "@commitlint/config-conventional": "17.0.0",
-    "@material-ui/core": "4.12.4",
-    "@material-ui/icons": "4.11.3",
-    "@material-ui/lab": "4.0.0-alpha.61",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.83",
+    "@mui/material": "^5.8.1",
     "@rollup/plugin-babel": "5.3.1",
     "@rollup/plugin-commonjs": "22.0.0",
     "@rollup/plugin-node-resolve": "13.3.0",
@@ -121,6 +121,7 @@
     "npm-run-all": "4.1.5",
     "pinst": "^3.0.0",
     "prettier": "^2.6.2",
+    "react": "18.1.0",
     "react-query": "3.39.0",
     "react-test-renderer": "17.0.2",
     "rimraf": "3.0.2",
